Fixed Application channels.
The code wasn't handling these channels at all - they are now read from the media server and stored like the other channel types. Also made the Applications item behave the same as the other channel items (hiding if none are available) and added support for the recently viewed list on the home screen.