Fix Server and Connection merging.
It turns out that doing operator overloading on classes that are
wrapped in shared_ptr’s are not a good idea. They will never be called
when you do if (sharedPtr == sharedPtr2) it will only test that the
reference is correct in both.

I have now removed the operator== on both PlexConnection and PlexServer
and implemented Equals() instead to use that. This has caused massive
problems with connection testing for a long time.