Disable matroska transcode streaming by default
Unfortunately it was hard to get XBMC’s code to cooperate with matroska
streaming. I could easily get it to request a new start point with
RestartMedia() but there where multiple issues with how it was reported
and how resuming would work. So now there is a new advanced
configuration to turn it on if you really want it
<usematroskatranscode>on</usematroskatranscode> in advancedsettings.xml